pyproject.toml
[project]
name = "phinder-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.1",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
